backend-python/requirements.txt
asgiref==3.8.1
bleach==6.2.0
Django==5.1.7
# django-filter==25.1
djangorestframework==3.15.2
dr-scaffold==2.1.2
drf-api-logger==1.1.16
# drf-yasg==1.21.10
inflect==7.5.0
inflection==0.5.1
isort==6.0.1
more-itertools==10.6.0
packaging==24.2
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
typeguard==4.4.2
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.1.1
webencodings==0.5.1

django-filter==24.3
drf-yasg==1.21.8